cli/jira_api/requirements.txt
atlassian-python-api
python-dotenv
pytest
